Refactor issues methods in RepositoriesController to ReposController
