+++ app/views/api/v1/repos/closed_issues.json.jbuilder
@@ -1,0 +1,1 @@
+json.partial! 'api/v1/repos/issues', issues: @closed_issues
