--- app/views/api/v1/repositories/open_issues.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! 'api/v1/repositories/issues', issues: @open_issues
