+++ spec/controllers/api/v1/repos_controller_spec.rb
@@ -1,10 +1,10 @@
 require 'rails_helper'
 
-RSpec.describe Api::V1::RepositoriesController, :vcr do
+RSpec.describe Api::V1::ReposController, :vcr do
 
   describe "GET#all_issues", type: :controller do
     before(:each) do
-      get :all_issues, organization_id: 'icalialabs', repository_id: 'watcher', format: :json
+      get :all_issues, id: 'icalialabs', repository_id: 'watcher', format: :json
     end
 
     it "returns the correct number of open and closed issues in watcher(13)" do
@@ -18,11 +18,11 @@ RSpec.describe Api::V1::RepositoriesController, :vcr do
 
   describe "GET#open_issues", type: :controller do
     before(:each) do
-      get :open_issues, organization_id: 'icalialabs', repository_id: 'watcher', format: :json
+      get :open_issues, id: 'icalialabs', repository_id: 'watcher', format: :json
     end
 
-    it "returns the correct number of open issues in watcher(4)" do
-      expect(assigns(:open_issues).count).to eq(4)
+    it "returns the correct number of open issues in watcher(3)" do
+      expect(assigns(:open_issues).count).to eq(3)
     end
 
     it "should be succesful" do
@@ -32,11 +32,11 @@ RSpec.describe Api::V1::RepositoriesController, :vcr do
 
   describe "GET#closed_issues", type: :controller do
     before(:each) do
-      get :closed_issues, organization_id: 'icalialabs', repository_id: 'watcher', format: :json
+      get :closed_issues, id: 'icalialabs', repository_id: 'watcher', format: :json
     end
 
-    it "returns the correct number of closed issues in watcher(9)" do
-      expect(assigns(:closed_issues).count).to eq(9)
+    it "returns the correct number of closed issues in watcher(10)" do
+      expect(assigns(:closed_issues).count).to eq(10)
     end
 
     it "should be succesful" do
